# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Docs api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/offers/": {
            "get": {
                "tags": [
                    "Offers"
                ],
                "summary": "Get product logs",
                "description": "Retrieves product logs based on query parameters.",
                "operationId": "getOffers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for product logs",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "seller_name",
                        "in": "query",
                        "description": "Seller name to filter by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country to filter by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Offers"
                ],
                "summary": "Create product logs",
                "description": "Creates new product logs based on provided data.",
                "operationId": "storeOffer",
                "requestBody": {
                    "description": "Product log data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "seller_id",
                                                "amount",
                                                "country"
                                            ],
                                            "properties": {
                                                "seller_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "US"
                                                },
                                                "archive": {
                                                    "type": "string",
                                                    "format": "binary"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data processed successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/banks": {
            "get": {
                "tags": [
                    "Banks"
                ],
                "summary": "Get bank logs",
                "description": "Retrieves bank logs based on query parameters.",
                "operationId": "getbanks",
                "parameters": [
                    {
                        "name": "bank_link",
                        "in": "query",
                        "description": "Bank link to filter by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "balance",
                        "in": "query",
                        "description": "Balance to filter by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BankLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Banks"
                ],
                "summary": "Create bank logs",
                "description": "Creates new bank logs based on provided data.",
                "operationId": "storeBankLog",
                "requestBody": {
                    "description": "Bank log data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "seller_id",
                                                "amount",
                                                "balance",
                                                "bank_link"
                                            ],
                                            "properties": {
                                                "seller_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "balance": {
                                                    "type": "string",
                                                    "example": "500"
                                                },
                                                "bank_link": {
                                                    "type": "string",
                                                    "example": "https://www.examplebank.com/user"
                                                },
                                                "archive": {
                                                    "type": "string",
                                                    "format": "binary"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "format": "binary"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data processed successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BankLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BankLog": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "seller_id": {
                        "type": "integer"
                    },
                    "amount": {
                        "type": "integer"
                    },
                    "balance": {
                        "type": "string"
                    },
                    "bank_link": {
                        "type": "string"
                    },
                    "archive_link": {
                        "type": "string"
                    },
                    "image_link": {
                        "type": "string"
                    },
                    "on_sale": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ProductLog": {
                "required": [
                    "seller_id",
                    "amount",
                    "archive_link",
                    "country"
                ],
                "properties": {
                    "seller_id": {
                        "description": "The ID of the seller associated with the product log",
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "description": "The amount or value associated with the product log",
                        "type": "integer",
                        "example": 100
                    },
                    "archive_link": {
                        "description": "URL to the archive containing product log details",
                        "type": "string",
                        "example": "http://example.com/archives/productlog1.zip"
                    },
                    "is_sold": {
                        "description": "Indicates whether the product log has been sold",
                        "type": "boolean",
                        "example": false
                    },
                    "country": {
                        "description": "The country associated with the product log",
                        "type": "string",
                        "example": "US"
                    },
                    "on_sale": {
                        "description": "Indicates whether the product log is available for sale",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}